==> 06D_clean_ppcs_2005.Rmd <==
---
title: "ppcs_2005"
output: html_document
---

## R Markdown

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

```{r 2005a }
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, foreign, here)

```


```{r 2005b }
# Loading in the dataset

ppcs_2005 <- read.spss(here("raw_data", "20020-0001-Data.sav"), to.data.frame=TRUE)
```


```{r 2005c }
#Renaming columns to see the people who had some sort of contact with the police
ppcs_2005 <- ppcs_2005 %>%
  mutate(face_to_face = V10) %>%
  mutate(contact = NA_real_) %>%
  mutate(num_face_to_face = NA_real_)
```


```{r }
# Civilian_race
ppcs_2005 <- ppcs_2005 %>% 
  mutate(civilian_race = case_when(
    (V5 == "White only" & V6 == "No") ~ "white", # white and not hispanic
    (V5 == "Black only") ~ "black", # black and black-hispanic
    (V6 == "Yes" & V5 != "Black only") ~ "hispanic", # hispanic
    TRUE ~ "other" # other
  ))
```


```{r cars}
# Civilian_age
ppcs_2005 <- ppcs_2005 %>% 
  mutate(civilian_age = V4)

```


```{r }
# Civilian_gender
ppcs_2005 <- ppcs_2005 %>%
  mutate(civilian_gender = case_when((V3 == "Male") ~ 1,
                                     (V3 == "Female") ~ 2
  ))

```


```{r }
# Civilian_income
ppcs_2005 <- ppcs_2005 %>%
  mutate(civilian_income = case_when((INCOME == "Less than $20,000 or na") ~ 1,
                                     (INCOME == "$20,000-$49,000") ~ 2,
                                     (INCOME == "$50,000 or more") ~ 3
    
  ))

```


```{r }
# Civilian_employed
ppcs_2005 <- ppcs_2005 %>%
  mutate(civilian_employed = case_when(
                              (WORK == "Yes") ~ 1,
                              TRUE ~ 0))

```


```{r }
# Population_size
ppcs_2005 <- ppcs_2005 %>%
  mutate(population_size = case_when((PLACE == "Under 100,000/not in a place") ~ 1,
                                     (PLACE == "100,000-499,999") ~ 2,
                                     (PLACE == "500,000-999,999") ~ 3,
                                     (PLACE == "1 million or more") ~ 4,
    
  ))

```


```{r }
# Time_of_encounter
ppcs_2005 <- ppcs_2005 %>%
  mutate(time_of_encounter = case_when(
                              V48 == "Yes" ~ 6,
                              V48 == "No" ~ 3,
                              TRUE ~ NA_real_
  ))
```


```{r }
# Officer_race
ppcs_2005 <- ppcs_2005 %>%
  mutate(off_black = case_when((V46 == "Black" | V47 == "All Black" | V47 == "Mostly Black") ~ 1,
                               TRUE ~ 0
    )) %>%
  mutate(off_white = case_when((V46 == "White" | V47 == "All White" | V47 == "Mostly White") ~ 1,
                               TRUE ~ 0
                               )) %>% 
  mutate(off_other = case_when((V46 == "Other" | V47 == "All some other race" | V47 == "Mostly some other race") ~ 1,
                               TRUE ~ 0
  )) %>%
  mutate(off_split = case_when(V47 == "Equally mixed" ~ 1,
                               TRUE ~ 0
  )) %>%
  mutate(off_hispanic = 0)
```


```{r}
# Type_of_incident
ppcs_2005 <- ppcs_2005 %>%
  mutate(type_of_incident = case_when((V27 == "Yes") ~ 2, 
                                      (V29 == "Yes") ~ 3,
                                      (V30 == "Yes") ~ 3,
                                      (V31 == "Yes") ~ 3,
                                      (V32 == "Yes") ~ 3,
                                      (V25 == "Yes") ~ 3,
                                      TRUE ~ NA_real_))
```


```{r there-is-not-civilian-behavior}
# Civilian_behavior
ppcs_2005 <- ppcs_2005 %>%
  mutate(civilian_behavior = 0)
```


```{r}
# Alternative_outcomes
ppcs_2005 <- ppcs_2005 %>%
  mutate(civilian_searched = case_when((V75 == "Yes" | V83 == "Yes") ~ 1,
                                       TRUE ~ 0)) %>%
  mutate(civilian_arrested = case_when((V61 == "Yes" | V82 == "Yes") ~ 1,
                                       TRUE ~ 0)) %>%
  mutate(civilian_guilty_of_illegal = case_when((V56 == "Yes" | V78 == "Yes") ~ 1,
                                       TRUE ~ 0)) %>%
  mutate(civilian_injured = case_when((INJURED == "Yes") ~ 1,
                                                TRUE ~ 0)) %>%
  mutate(excess_force = case_when((EXFORCE == "Yes") ~ 1,
                                      TRUE ~ 0))
```


```{r}
#Use of force 
ppcs_2005 <- ppcs_2005 %>%
  mutate(any_force = case_when(
    # (V12 == "Yes") ~ 1,
    (V14 == "Yes") ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(grab_push = case_when(
    (V17 == 'Push or grab') ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(hit_kick = case_when(
    (V18 == 'Kick or hit') ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(point_gun = case_when(
    (V20 == 'Point gun') ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(handcuffed = case_when(
    (FRCHCFF == 'Yes') ~ 1,
    (V81 == 'Yes') ~ 1,
    (V60 == 'Yes') ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(pepper_stun = case_when(
    (V19 == "Use pepper spray") ~ 1,
    TRUE ~ 0
  ))
```


```{r }
#add year column
ppcs_2005 <- ppcs_2005 %>%
  mutate(year = 2005)

```


```{r }
# final dataframe
ppcs_2005 <- ppcs_2005 %>%
  select(civilian_race,
         civilian_age,
         civilian_gender,
         civilian_income,
         civilian_employed,
         population_size,
         time_of_encounter,
         off_black,
         off_white,
         off_hispanic,
         off_other,
         off_split,
         type_of_incident,
         civilian_behavior,
         civilian_searched,
         civilian_arrested,
         civilian_guilty_of_illegal,
         civilian_injured,
         excess_force,
         year,
         contact,
         face_to_face,
         num_face_to_face,
         any_force,
         grab_push, 
         hit_kick,
         point_gun,
         handcuffed,
         pepper_stun
         )
```


```{r rdata file}
save(ppcs_2005, file = 'clean_data/ppcs_2005.RData')
sessionInfo()
```
## Including Plots

You can also embed plots, for example:

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
