==> 06E_clean_ppcs_2008.Rmd <==
---
title: "ppcs 2008"
output: html_document
---

```{r}
# load 2008 ppcs data

if (!require("pacman")) install.packages("pacman")
pacman::p_load(readr, tidyverse, here)
```

```{r}
ppcs_2008 <- read_tsv(here("raw_data","32022-0001-Data.tsv"))

```

```{r}
ppcs_2008 <- ppcs_2008 %>%
  mutate(face_to_face = V2) %>%
  mutate(contact = NA_real_) %>%
  mutate(num_face_to_face = NA_real_)
```

```{r}
#Civilian Race
ppcs_2008 <- ppcs_2008 %>% 
  mutate(civilian_race = case_when(
    (RACE == 1 & HISP == 2) ~ "white",
    (RACE == 2 ) ~ "black",
    (HISP == 1 & RACE != 2) ~ "hispanic",
    TRUE ~ "other"
  ))

#check the race distribution
summary(ppcs_2008$civilian_race)
```

```{r}
#Civilian Age
ppcs_2008 <- ppcs_2008 %>%
  mutate( civilian_age = AGE)
```

```{r}
#Civilian Gender
ppcs_2008 <- ppcs_2008 %>%
  mutate(civilian_gender = SEX)
```

```{r}
#Civilian Income
# missing column in 2008
ppcs_2008 <- ppcs_2008 %>%
  mutate(civilian_income = NA)
```

```{r}
#Civilian Employed
#since civilian employed (according to appendix) is coded as only 1(yes) or 0(otherwise),
ppcs_2008 <- ppcs_2008 %>%
  mutate(civilian_employed = ifelse(WORK != 1, 0, 1))
```

```{r}
#Population Size
# missing column in 2008, make all 1,

ppcs_2008 <- ppcs_2008 %>%
  mutate(population_size = NA)
```

```{r}
#Time of encounter
#Time of encounter: appendix does not include 0, 8, 9 so will make those NA
day <- c(1:3)
night <- c(4:6)
ppcs_2008 <- ppcs_2008 %>% 
  mutate(time_of_encounter = case_when(
    (V4 %in% day) ~ 3,
    (V4 %in% night) ~ 6,
    TRUE ~ NA_real_
  )) %>%
  mutate(time_of_encounter = as.factor(time_of_encounter))

summary(ppcs_2008$time_of_encounter)
```

```{r}
#Officer Race
# create 3 separate columns for each dummy variable
#check if multiple officers (V24B) race has value or
# if only 1 officer race(v24A) has
# equally mixed was calculated for both black and white

#officer race black:
# includes:
# multiple officers (V24B) black, mostly black, equally mixed
ppcs_2008 <- ppcs_2008 %>%
  mutate(off_black = case_when(
    (V24B == 2 | V24B == 5 | V24A == 2 ) ~ 1,
    TRUE ~ 0
  ))

#officer race white:
#includes:
#multiple officers (V24B) white, mostly white, equally mixed
ppcs_2008 <- ppcs_2008 %>%
  mutate(off_white = case_when(
    (V24B == 1 | V24B == 4 | V24A == 1 ) ~ 1,
    TRUE ~ 0
  ))

#officer race other
#includes:
#multiple officers (V24B) 
ppcs_2008 <- ppcs_2008 %>%
  mutate(off_other = case_when(
    (V24B == 3 | V24B == 6 | V24A == 3 ) ~ 1,
    TRUE ~ 0
  ))
```

```{r}
#Officer Split Race
ppcs_2008 <- ppcs_2008 %>%
  mutate(off_split = case_when(
    (V24B == 7)~ 1,
    TRUE ~ 0
  ))
```

```{r}
#Officer Hispanic Race
ppcs_2008 <- ppcs_2008 %>%
  mutate(off_hispanic = 0)

```

```{r}
#Type of Incident
ppcs_2008 <- ppcs_2008 %>%
  mutate(type_of_incident = case_when(
    (REASON %in% 2:3) ~ 2,
    (REASON %in% c(1, 4:7)) ~ 3,
    TRUE ~ NA_real_
))

```

```{r}
#Civilian Behavior
# 1 if any of the columns are 1
# 0 if all are 0
ppcs_2008 <- ppcs_2008 %>%
  mutate(civilian_behavior = case_when(
    (V12A == 1 | V12B == 1 | V12C == 1 | V12D == 1 | V11 == 1 | V12E == 1) ~ 1,
    TRUE ~ 0
))

```

```{r}
#Alternative Outcomes
# searched 1 if yes and 0 otherwise
ppcs_2008 <- ppcs_2008 %>% 
  mutate(civilian_searched = ifelse(SEARCH == 1, 1, 0 ))

#arrested
ppcs_2008 <- ppcs_2008 %>% 
  mutate(civilian_arrested = ifelse(V9 == 1, 1, 0 ))

#found 
ppcs_2008 <- ppcs_2008 %>%
  mutate(civilian_guilty_of_illegal = ifelse(FOUND == 1, 1, 0 ))

#injured
ppcs_2008 <- ppcs_2008 %>%
  mutate(civilian_injured = ifelse(V8 == 1, 1, 0))

#excessive force
ppcs_2008 <- ppcs_2008 %>%
  mutate(excess_force = ifelse(V7 == 1, 1, 0))

```

```{r}
#Force Use
ppcs_2008 <- ppcs_2008 %>%
  mutate(any_force = case_when(
    (V5 == 1) ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(grab_push = case_when(
    (V6D == 1) ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(hit_kick = case_when(
    (V6E == 1) ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(point_gun = case_when(
    (V6H == 1) ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(handcuffed = case_when(
    (V10 == 1) ~ 1,
    TRUE ~ 0
  )) %>%
  mutate(pepper_stun = case_when(
    (V6F == 1) ~ 1,
    (V6G == 1) ~ 1,
    TRUE ~ 0
  ))

```

```{r}
#Cleaned and selected columns
ppcs_2008 <- ppcs_2008 %>%
  select(civilian_race, 
         civilian_age, 
         civilian_gender, 
         civilian_income, 
         civilian_employed, 
         population_size, 
         time_of_encounter, 
         off_black, 
         off_white, 
         off_other, 
         off_split, 
         off_hispanic, 
         type_of_incident, 
         civilian_behavior, 
         civilian_searched, 
         civilian_arrested, 
         civilian_guilty_of_illegal, 
         civilian_injured, 
         excess_force, 
         contact, 
         face_to_face, 
         num_face_to_face, 
         any_force,
         grab_push, 
         hit_kick,
         point_gun,
         handcuffed,
         pepper_stun) %>%
  mutate(year = 2008)

save(ppcs_2008, file = "clean_data/ppcs_2008.Rdata")
```

```{r}
sessionInfo()
```
