==> 06F_clean_ppcs_2011.Rmd <==
---
title: "ppcs_2011"
output: html_document
--

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, here)

load(here("raw_data", "34276-0001-Data.rda"))
ppcs_2011 <- da34276.0001
```



```{r}
#RACE
ppcs_2011 <- ppcs_2011 %>%
  mutate(civilian_race =  case_when((RACE == '(01) White Only' & 
                                       HISP != '(1) Respondent is Hispanic') ~ 'white',
                                    (RACE == '(02) Black Only') ~ 'black',
                                    (HISP == '(1) Respondent is Hispanic' & 
                                       RACE != '(02) Black Only') ~ 'hispanic',
                                    TRUE ~ 'other'))

```




```{r}
#AGE                                 
ppcs_2011 <- ppcs_2011 %>% 
  mutate(civilian_age = AGE) 
```




```{r}
#INCOME
ppcs_2011 <- ppcs_2011 %>% 
  mutate(civilian_income = case_when(INCOME =='(1) Less than $20,000 or NA' ~ 1, 
                                     INCOME == '(2) $20,000 - $49,999' ~ 2,
                                     INCOME == '(3) $50,000 or more' ~ 3)) 

```




```{r}
#GENDER
ppcs_2011 <- ppcs_2011 %>% 
  mutate(civilian_gender = case_when(
    (SEX == '(1) Male') ~ 1,
    (SEX == '(2) Female') ~ 2,
    TRUE ~ NA_real_
))
```


```{r}
#EMPLOYED
ppcs_2011 <- ppcs_2011 %>% 
  mutate(civilian_employed = case_when(WORK == '(1) Yes' ~ 1,
                                      WORK == '(2) No' ~ 0,
                                      TRUE ~ 0)) 
```


```{r}
#POPULATION
ppcs_2011 <- ppcs_2011 %>%
  mutate(population_size = case_when(PLACE == '(1) Under 100,000/Not in a place' ~ 1, 
                                     PLACE == '(2) 100,000-499,999' ~ 2,
                                     PLACE == '(3) 500,000-999,999' ~ 3,
                                     PLACE == '(4) 1 million or more' ~ 4)) 
```


```{r}
#TIME OF ENCOUNTER
ppcs_2011 <- ppcs_2011 %>% 
 mutate(time_of_encounter = case_when(
   (V81 == ('(01) After 6 a.m. - 12 noon')| 
      V204 == ('(01) After 6 a.m. - 12 noon')) ~3,
   (V81 == ('(02) After 12 noon - 6 p.m.')|
      V204 == ('(02) After 12 noon - 6 p.m.')) ~3, 
    (V81 == ('(01) After 6 a.m. - 12 noon')| 
       V204 == ('(01) After 6 a.m. - 12 noon'))~3,
     (V81 == ('(04) After 6 p.m. - 12 midnight')| 
        V204 == ('(04) After 6 p.m. - 12    midnight')) ~ 6,
      (V81 == ('(05) After 12 midnight - 6 a.m.')| 
         V204 == ('(05) After 12 midnight - 6 a.m.')) ~ 6,
       (V81 == ("(06) Don't know what time of night")| 
          V204 == ("(06) Don't know what time of night")) ~ 6,
        TRUE ~ NA_real_))
```


```{r}
#OFFICER RACE 
ppcs_2011 <- ppcs_2011 %>%
  mutate(off_black = case_when(((V69 == '(1) Black or African American') | 
                                  (V76 == '(1) Black or African American') | 
                                  (V192 == '(1) Black or African American') |
                                  (V199 == '(1) Black or African American')) ~ 1, 
                                   TRUE ~ 0)) %>%
  mutate(off_white = case_when(((V68 == '(1) White') | 
                                  (V75 == '(1) White') | 
                                  (V191 == '(1) White') | 
                                  (V198 == '(1) White'))~ 1, 
                                   TRUE ~ 0)) %>%
  mutate(off_hispanic = case_when(((V67 == '(1) Yes') | 
                                     (V74 == '(1) Yes') |
                                    (V190 == '(1) Yes') | 
                                     (V197 == '(1) Yes')) ~ 1,  
                                    TRUE ~ 0)) %>%
  mutate(off_other = case_when(
    (V70 == '(1) American Indian or Alaska Native' | 
      V77 == '(1) American Indian or Alaska Native' |  
      V71 == '(1) Asian'|
    V78 == '(1) Asian' | 
      V72 == '(1) Native Hawaiian or other Pacific Islander' |
    V79 == '(1) Native Hawaiian or other Pacific Islander' | 
      V193 == '(1) American Indian or Alaska Native' | 
      V200 == '(1) American Indian or Alaska Native' | 
      V194 == '(1) Asian'| 
      V201 == '(1) Asian'| 
      V195 == '(1) Native Hawaiian or other Pacific Islander' | 
      V202 == '(1) Native Hawaiian or other Pacific Islander')~ 1,
       TRUE ~ 0))
```


```{r}
#CIVILIAN BEHAVIOR

ppcs_2011 <- ppcs_2011 %>%
  mutate(civilian_behavior = case_when((V164 == '(1) Yes') | (V306 == '(1) Yes')~ 1,
                                       (V166 == '(1) Yes')| (V308 == '(1) Yes') ~ 1,
                                       (V170 == '(1) Yes')| (V312 == '(1) Yes') ~ 1,
                                       (V172 == '(1) Yes')| (V314 == '(1) Yes') ~ 1,
                                       (V174 == '(1) Yes') | (V316 == '(1) Yes') ~ 1,
                                       (V176 == '(1) Yes') | (V318 == '(1) Yes') ~ 1,
                                       (V178 == '(1) Yes')| (V320 == '(1) Yes') ~ 1,
                                       TRUE ~ 0)) 
```


```{r}
#ALTERNATIVE OUTCOMES
ppcs_2011 <- ppcs_2011 %>%
  
  #searched
  mutate(civilian_searched = case_when(
    
      (V183 == '(1) Yes' | V327 == 	'(1) Yes') ~ 1, 
      TRUE ~ 0)) %>%
  
  #arrested
  mutate(civilian_arrested = case_when(
      (V182 == '(1) Yes' | V324 == '(1) Yes') ~ 1, 
        TRUE ~ 0)) %>%
  
  #civilian guilty if carrying illegal items
  mutate(civilian_guilty_of_illegal = case_when(
      (V187 == '(1) Yes' | V334 == '(1) Yes') ~ 1, 
        TRUE ~ 0)) %>%
  
  #civilian injured
  mutate(civilian_injured = case_when(
      (V180 == '(1) Yes' | V322 == '(1) Yes')~ 1,
      TRUE ~ 0)) %>%
  
  #excess force
  mutate(excess_force = case_when(
      (V163 == '(1) Yes' | V305 == '(1) Yes' | V378 == '(1) Yes') ~1,
      TRUE ~ 0))


```


```{r}
#TYPE OF INCIDENT
ppcs_2011 <- ppcs_2011 %>% 
   mutate(type_of_incident = case_when(
                                       (V11 == '(1) Yes') ~ 1,
                                       (!is.na(V34)) ~ 1,
                                       (V13 == '(1) Yes') ~ 2,
                                       (V15 == '(1) Yes') ~ 2,
                                       (V7 == '(1) Yes') ~ 3,
                                       #(V17 == '(1) Yes') ~ 3,
                                       TRUE ~ NA_real_))
```


```{r}
#OFFICER SPLIT
ppcs_2011 <- ppcs_2011 %>%
  mutate(off_split = 0)

```

```{r}
#CONTACT, Number of Face to Face contacts, Face to face
ppcs_2011 <- ppcs_2011 %>%
  mutate(contact = NA) %>%
  mutate(num_face_to_face = V30) %>%
  mutate(face_to_face = V31)
```

```{r}
#FORCE USE 
ppcs_2011 <- ppcs_2011 %>% 
  mutate(any_force = case_when((V160 == '(1) Yes') ~ 1,
    (V301 == '(1) Yes') ~ 1,
    TRUE ~ 0)) %>%
  mutate(grab_push = case_when((V148 == '(1) Yes') ~ 1, 
    (V289 == '(1) Yes') ~ 1, TRUE ~ 0)) %>%
  mutate(hit_kick = case_when((V152 == '(1) Yes') ~ 1,
    (V293 == '(1) Yes') ~ 1, TRUE ~ 0)) %>%
  mutate(point_gun = case_when( (V158 == '(1) Yes') ~ 1,
    (V299 == '(1) Yes') ~ 1, TRUE ~ 0)) %>%
  mutate(handcuffed = case_when((V150 == '(1) Yes') ~ 1,
    (V291 == '(1) Yes') ~ 1, TRUE ~ 0)) %>%
  mutate(pepper_stun = case_when((V154 == '(1) Yes') ~ 1,
    (V156 == '(1) Yes') ~ 1,
    (V295 == '(1) Yes') ~ 1,
    (V297 == '(1) Yes') ~ 1,
    TRUE ~ 0))
```


```{r}
#YEAR
ppcs_2011 <- ppcs_2011 %>%
  mutate(year = 2011) 
```


```{r}
ppcs_2011 <- ppcs_2011 %>%
  select(civilian_race,
         civilian_age,
         civilian_gender,
         civilian_income,
         civilian_employed,
         population_size,
         time_of_encounter,
         off_black,
         off_white,
         off_hispanic,
         off_other,
         type_of_incident,
         civilian_behavior,
         civilian_searched,
         civilian_arrested,
         civilian_guilty_of_illegal,
         civilian_injured,
         excess_force,
         off_split,
         year,
         any_force,
         grab_push,
         hit_kick,
         point_gun,
         handcuffed,
         pepper_stun,
         contact,
         num_face_to_face,
         face_to_face
         )

save(ppcs_2011, file = "clean_data/ppcs_2011.RData")
```


```{r}
sessionInfo()
```
