==> 09_ppcs_summary_stats.Rmd <==
---
title: "replicating_ppcs"
author: "Brenda Fried"
date: "7/18/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load Libraries
```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(ggplot2, scales, here, tidyverse)
```

```{r load-data}
load(here("clean_data", "merged_ppcs.RData"))

```
# Proportion of Race
```{r}
#percentage of races from whole sample
merged_ppcs %>%
  group_by(civilian_race) %>%
  summarize(count = n()) %>% 
  mutate(prop = count/sum(count))
```
#Proportions of gender 
```{r}
#percentage of male and female from whole sample
merged_ppcs %>%
  group_by(civilian_gender) %>%
  summarize(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r}
#percentage of gender per race
merged_ppcs %>%
  group_by(civilian_race, civilian_gender) %>%
  summarize(gender_count = n()) %>%
  ungroup() %>%
  group_by(civilian_race) %>%
  mutate(sum = sum(gender_count)) %>%
  mutate(prop = gender_count/sum)

```
# Average age
```{r avg-age-full-sample}
merged_ppcs %>%
  select(civilian_age) %>% 
  filter(civilian_age != 'missing/ni/ou') %>%
  mutate(civilian_age = as.numeric(civilian_age)) %>%
  summarize(mean = mean(civilian_age))
```
```{r avg-age-per-race}
merged_ppcs %>%
  filter(civilian_age != 'missing/ni/ou') %>%
  mutate(civilian_age = as.numeric(civilian_age)) %>%
  group_by(civilian_race) %>%
  summarize(mean = mean(civilian_age))
```

# Proportion of Civilian Employed
```{r employed-full-sample}
merged_ppcs %>%
  group_by(civilian_employed) %>%
  summarize(count = n()) %>%
  mutate(prop = count/sum(count))

```


```{r employed-per-race}
merged_ppcs %>%
  mutate(civilian_employed = as.numeric(civilian_employed)) %>%
  group_by(civilian_race, civilian_employed) %>%
  summarize(employed_count = n()) %>% 
  ungroup() %>%
  group_by(civilian_race) %>%
  mutate(sum = sum(employed_count)) %>%
  filter(civilian_employed == 1) %>%
  mutate(prop = employed_count/sum)
  
```
# Average Income as factor
```{r income-full-sample}
merged_ppcs %>%
  summarise(mean = mean(as.numeric(civilian_income), na.rm = TRUE))
  
```


```{r income-by-race}
merged_ppcs %>%
  group_by(civilian_race) %>%
  summarise(mean = mean(as.numeric(civilian_income), na.rm = TRUE))
```
# Average Population Size

```{r pop-size-full-sample}
merged_ppcs %>%
  summarise(mean = mean(as.numeric(population_size), na.rm = TRUE))
```

```{r pop-size-by-race}
merged_ppcs %>%
  group_by(civilian_race) %>%
  summarise(mean = mean(as.numeric(population_size), na.rm = TRUE))
```

# Proportion of Incident Type
```{r incident-type-full-sample}
merged_ppcs %>%
  filter(!is.na(type_of_incident)) %>%
  group_by(type_of_incident) %>%
  summarize(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r incident-type-by-race}
merged_ppcs %>%
  filter(is.na(type_of_incident) != TRUE) %>%
  mutate(type_of_incident = as.factor(type_of_incident)) %>%
  group_by(civilian_race, type_of_incident) %>%
  summarise(c = n()) %>%
  ungroup() %>%
  group_by(civilian_race) %>%
  mutate(sum = sum(c)) %>%
  mutate(prop = c/sum)
```

# Proportion of Time of Encounter
```{r time-of-encounter-full-sample}
merged_ppcs %>%
  group_by(time_of_encounter) %>%
  filter(is.na(time_of_encounter) != TRUE ) %>%
  summarise(count = n()) %>%
  ungroup() %>%
  mutate(prop = count/ sum(count))


```

```{r time-of-encounter-by-race}
merged_ppcs %>%
  filter(is.na(time_of_encounter) != TRUE ) %>%
  group_by(civilian_race, time_of_encounter) %>%
  summarize(enc_count = n()) %>% 
  ungroup() %>%
  group_by(civilian_race) %>%
  mutate(sum = sum(enc_count)) %>%
  mutate(prop = enc_count/sum)

```
# Proportion of Civilian Behavior 
(not separated - used as one column because regression uses it together)
```{r civilian-behavior-full-sample}
merged_ppcs %>%
  group_by(civilian_behavior) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
  
```

```{r civilian-behavior-by-race}
merged_ppcs %>%
  group_by(civilian_race, civilian_behavior) %>%
  summarize(beh_count = n()) %>% 
  ungroup() %>%
  group_by(civilian_race) %>%
  mutate(sum = sum(beh_count)) %>%
  mutate(prop = beh_count/sum) %>% 
  filter(civilian_behavior == 1) %>%
  select(civilian_race, civilian_behavior, prop)
```

# Proportion of Officer Race
```{r officer-race-full-sample}

merged_ppcs %>%
  mutate(off_race = case_when(
    (off_black == 1) ~ 'black',
    (off_white == 1) ~ 'white',
    (off_other == 1) ~ 'other',
    (off_split == 1) ~ 'split',
    (off_hispanic == 1)~ 'hispanic'
  )) %>%
  filter(is.na(off_race) == FALSE) %>%
  group_by(off_race) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r}
merged_ppcs %>%
  mutate(off_race = case_when(
    (off_black == 1) ~ 'black',
    (off_white == 1) ~ 'white',
    (off_other == 1) ~ 'other',
    (off_split == 1) ~ 'split',
    (off_hispanic == 1)~ 'hispanic'
  )) %>%
  filter(is.na(off_race) == FALSE) %>%
  group_by(civilian_race, off_race) %>%
  summarise(offrace_count = n()) %>%
  ungroup() %>%
  group_by(civilian_race) %>%
  mutate(sum = sum(offrace_count)) %>%
  mutate(prop = offrace_count/sum) %>%
  select(civilian_race, off_race, prop)

```

# Proportion of Alternative outcomes

```{r searched-full-sample}
merged_ppcs %>%
  group_by(civilian_searched) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r searched-by-race}
merged_ppcs %>%
  group_by(civilian_race, civilian_searched) %>%
  summarise(search_count = n()) %>%
  mutate(prop = search_count/sum(search_count)) %>%
  filter(civilian_searched == 1)
```

# Proportion of arrested
```{r arrested-full-sample}
merged_ppcs %>%
  group_by(civilian_arrested) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```


```{r arrested-by-race}
merged_ppcs %>%
  group_by(civilian_race) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, civilian_arrested) %>%
  summarise(arrest_count = n()) %>%
  mutate(prop = arrest_count/sum(arrest_count)) %>%
  filter(civilian_arrested == 1)

```

# Proportion of civilians carrying illegal items
```{r guily-carrying-illegal-full-sample}
merged_ppcs %>%
  group_by(civilian_guilty_of_illegal) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r guilty-carrying-illegal-by-race}
merged_ppcs %>%
  group_by(civilian_race) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, civilian_guilty_of_illegal) %>%
  summarise(guilt_count = n()) %>%
  mutate(prop = guilt_count/sum(guilt_count)) %>%
  filter(civilian_guilty_of_illegal == 1)
```

# Proportion of civilian injured
```{r injured-full-sample}
merged_ppcs %>%
  group_by(civilian_injured) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r injured-by-race}
merged_ppcs %>%
  group_by(civilian_race, civilian_injured) %>%
  summarise(inj_count = n()) %>%
  mutate(prop = inj_count/sum(inj_count)) %>%
  filter(civilian_injured == 1)
```

# Proportion of perceiving excess force
```{r exess-force-full-sample}
merged_ppcs %>%
  group_by(excess_force) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```


```{r excess-force-by-race}
merged_ppcs %>%
  group_by(civilian_race, excess_force) %>%
  summarise(ex_count = n()) %>%
  mutate(prop = ex_count/sum(ex_count)) %>%
  filter(excess_force == 1)
```

#Check use of force

```{r any-force-full-sample} 
merged_ppcs %>%
  group_by(any_force) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r any-force-by-race}
merged_ppcs %>%
  group_by(any_force) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, any_force) %>%
  summarise(any_force_count = n()) %>%
  mutate(prop = any_force_count/sum(any_force_count)) %>%
  filter(any_force == 1)
```

```{r grab-push-full-sample}
merged_ppcs %>%
  group_by(grab_push) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r grab-push-by-race}
merged_ppcs %>%
  group_by(grab_push) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, grab_push) %>%
  summarise(grab_push_count = n()) %>%
  mutate(prop = grab_push_count/sum(grab_push_count)) %>%
  filter(grab_push == 1)
```

#hit and kick

```{r hit-kick-full-sample}
merged_ppcs %>%
  group_by(hit_kick) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r hit-kick-by-race}
merged_ppcs %>%
  group_by(hit_kick) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, hit_kick) %>%
  summarise(hit_kick_count = n()) %>%
  mutate(prop = hit_kick_count/sum(hit_kick_count)) %>%
  filter(hit_kick == 1)
```

#point gun

```{r point-gun-full-sample}
merged_ppcs %>%
  group_by(hit_kick) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r point-gun-by-race}
merged_ppcs %>%
  group_by(point_gun) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, point_gun) %>%
  summarise(point_gun_count = n()) %>%
  mutate(prop = point_gun_count/sum(point_gun_count)) %>%
  filter(point_gun == 1)
```

# Proportion of handcuffed

```{r handcuffed-full-sample}
merged_ppcs %>%
  group_by(handcuffed) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r handcuffed-by-race}
merged_ppcs %>%
  group_by(handcuffed) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, handcuffed) %>%
  summarise(handcuffed_count = n()) %>%
  mutate(prop = handcuffed_count/sum(handcuffed_count)) %>%
  filter(handcuffed == 1)
```
#pepperspray stungun

```{r pepper-stun-full-sample}
merged_ppcs %>%
  group_by(pepper_stun) %>%
  summarise(count = n()) %>%
  mutate(prop = count/sum(count))
```

```{r pepper-stun-by-race}
merged_ppcs %>%
  group_by(pepper_stun) %>%
  mutate(count = n()) %>%
  ungroup() %>%
  group_by(civilian_race, pepper_stun) %>%
  summarise(pepper_count = n()) %>%
  mutate(prop = pepper_count/sum(pepper_count)) %>%
  filter(pepper_stun == 1)
```

# Add at least use of force levels and factor
```{r}
merged_ppcs <- merged_ppcs %>%
  mutate(
    at_least_grabbed = ifelse(
      grab_push == 1 |
      handcuffed == 1|
      point_gun == 1 |
      hit_kick == 1  |
      pepper_stun == 1, 1, 0),
    
    at_least_handcuffed = ifelse(
      handcuffed == 1|
      point_gun == 1 |
      hit_kick == 1  |
      pepper_stun == 1, 1, 0),
    
    at_least_point_gun = ifelse(
      point_gun == 1 |
      hit_kick == 1  |
      pepper_stun == 1, 1, 0),
    
    at_least_kick_pepper_stun = ifelse(
      point_gun == 1 |
      hit_kick == 1  |
      pepper_stun == 1, 1, 0),
    
    force = ifelse(
      any_force == 1 |
      grab_push == 1 |
      handcuffed == 1|
      point_gun == 1 |
      hit_kick == 1  |
      pepper_stun == 1, 1, 0),
    year = as.factor(year),
    civilian_race = factor(civilian_race, levels = c("white", "black", "hispanic", "other")),
    civilian_gender = as.factor(civilian_gender),
    civilian_income = as.factor(civilian_income),
    civilian_employed = as.factor(civilian_employed),
    population_size = as.factor(population_size),
    time_of_encounter = as.factor(time_of_encounter),
    off_black = as.factor(off_black),
    off_white = as.factor(off_white),
    off_hispanic = as.factor(off_hispanic),
    off_other = as.factor(off_other),
    off_split = as.factor(off_split),
    type_of_incident = as.factor(type_of_incident),
    civilian_behavior = as.factor(civilian_behavior),
    civilian_searched = as.factor(civilian_searched),
    civilian_arrested = as.factor(civilian_arrested),
    civilian_guilty_of_illegal = as.factor(civilian_guilty_of_illegal),
    civilian_injured = as.factor(civilian_injured),
    excess_force = as.factor(excess_force),
    any_force = as.factor(any_force),
    grab_push = as.factor(grab_push),
    hit_kick = as.factor(hit_kick),
    point_gun = as.factor(point_gun),
    handcuffed = as.factor(handcuffed),
    pepper_stun = as.factor(pepper_stun),
    at_least_grabbed = as.factor(at_least_grabbed),
    at_least_handcuffed = as.factor(at_least_handcuffed),
    at_least_point_gun = as.factor(at_least_point_gun),
    at_least_kick_pepper_stun = as.factor(at_least_kick_pepper_stun),
    force = as.factor(force)
  ) 
save(merged_ppcs, file = 'clean_data/summarized_ppcs.RData')
sessionInfo()
```
