==> 10_c_sqf_tables.Rmd <==
---
title: "Stop, Question, and Frisk Logistic Regressions"
author: "Adnan Hoq"
date: "12/16/2019"
output:
  html_document: default
---

```{r setup, include=FALSE}
knitr::knit_meta(class=NULL, clean = TRUE)
knitr::opts_chunk$set(echo = TRUE)
```


```{r include=FALSE}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, gridExtra, naniar, scales, miceadds, here, data.table, knitr, kableExtra)

```

## Load the 2003-2013 SQF dataset

```{r load-data}

log_data <- readRDS('log_data2.rds')

```



## Run Our Logistic Regression Models

```{r log-model1, warning = FALSE}

#model with no control: race as the only predictor and Use of force as the outcome
model_no_cntrl <- glm(any_force_used ~ race,
                       data = log_data,
                       family = "binomial", y=FALSE, model=FALSE)


#create a data frame with the odds ratios of each race relative to the white mean from the no control model
no_control <- data.frame(Model= "No Control", WhiteMean = as.character(0.161),
                          Black = exp(coef(model_no_cntrl))[2], Hispanic = exp(coef(model_no_cntrl))[3],
                          Asian = exp(coef(model_no_cntrl))[4], Others = exp(coef(model_no_cntrl))[5])

summary(model_no_cntrl)
exp(coef(model_no_cntrl))

```

```{r log-model2, warning = FALSE}
#model with civilian demographics added as control
 model_civilian_demo <- glm(any_force_used ~ race + sex + age + I(age^2),
                            data = log_data,
                            family = "binomial", y=FALSE, model=FALSE)

#create a data frame with the odds ratios of each race relative to the white mean from the civilian demographics model
 
 civilian_demographics <- data.frame(Model = "(+) Civilian Demographics", WhiteMean = "",
                                     Black = exp(coef(model_civilian_demo))[2], Hispanic = exp(coef(model_civilian_demo))[3],
                                     Asian = exp(coef(model_civilian_demo))[4], Others = exp(coef(model_civilian_demo))[5])

summary(model_civilian_demo)
exp(coef(model_civilian_demo))
 
```



```{r log-model3, warning = FALSE}

#model with encounter characteristics added as control
 model_encnt_char <- glm(any_force_used ~ race + sex + I(age^2)
                         + inout + daytime + ac_incid + ac_time +
                           offunif + typeofid + othpers,
                         data = log_data,
                         family = "binomial", y=FALSE, model=FALSE)

 
#create a data frame with the odds ratios of each race relative to the white mean from the encounter characteristics model
 encounter_characteristics <- data.frame(Model = "(+) Encounter Characteristics", WhiteMean = "",
                                         Black = exp(coef(model_encnt_char))[2],
                                         Hispanic = exp(coef(model_encnt_char))[3],
                                         Asian = exp(coef(model_encnt_char))[4],
                                         Others = exp(coef(model_encnt_char))[5])




summary(model_encnt_char)
exp(coef(model_encnt_char))

```

```{r log-model4, warning = FALSE}
#model with civilian behavior added as control
 model_civ_behv <- glm(any_force_used ~ race + sex + I(age^2) + inout + daytime + ac_incid + ac_time +
                         offunif + typeofid + othpers + cs_bulge + cs_cloth + cs_casng + cs_lkout +
                         cs_descr + cs_drgtr + cs_furtv + cs_vcrim +
                         cs_objcs + cs_other + wepnfnd,
                       data = log_data,
                       family = "binomial", y=FALSE, model=FALSE)

#create a data frame with the odds ratios of each race relative to the white mean from the civilian behavior model
 civilian_behavior <- data.frame(Model = "(+) Civilian Behavior", WhiteMean = "", Black = exp(coef(model_civ_behv))[2],
                                 Hispanic = exp(coef(model_civ_behv))[3],
                                 Asian = exp(coef(model_civ_behv))[4],
                                 Others = exp(coef(model_civ_behv))[5])
summary(model_civ_behv)
exp(coef(model_civ_behv))

```


```{r log-model5, warning = FALSE}


#model with precinct and year levels added as controls
 load("model.rda")


#create a data frame with the odds ratios of each race relative to the white mean from the full control model
 full_control <- data.frame(Model = "(+) Precinct, Year", WhiteMean = "", Black = exp(coef(model_full_control))[2],
                            Hispanic = exp(coef(model_full_control))[3],
                            Asian = exp(coef(model_full_control))[4],
                            Others = exp(coef(model_full_control))[5])




#creating a dataframe with our results
 our_results <- bind_rows(no_control, civilian_demographics, encounter_characteristics,
                          civilian_behavior, full_control) %>%
   mutate(Black = round(Black, 3),
          Hispanic = round(Hispanic,3), Asian = round(Asian, 3), Others = round(Others, 3))


#assigning new row names to the "our_results" data frame
 rownames(our_results) <- c("(a)", "(b)",
                            "(c)",
                            "(d)", "(e)")
 
summary(model_full_control)
exp(coef(model_full_control))
```


## Results of Fryer's Logistic Regression Models

```{r fryer-model, warning = FALSE}


#Fryer's no control model
fryer_no_control <- data.frame(Model = "No Control", WhiteMean = as.character(0.153),Black = 1.534, Hispanic = 1.582, Asian = 1.044,
                                Others = 1.392)



#Fryer's model with civilian demographics added as control
fryer_civilian_demographics <- data.frame(Model = "(+) Civilian Demographics", WhiteMean = "", Black = 1.480, Hispanic = 1.517,
                                          Asian = 1.010,
                                           Others = 1.346)


#Fryer's model with encounter characteristics added as control
fryer_encounter_characteristics <- data.frame(Model = "(+) Encounter Characteristics", WhiteMean = "", Black = 1.655, Hispanic = 1.641,
                                              Asian = 1.059,
                                               Others = 1.452)

#Fryer's model with civilian behavior added as control
fryer_civilian_behavior <- data.frame(Model = "(+) Civilian Behavior", WhiteMean = "", Black = 1.462, Hispanic = 1.516,
                                       Asian = 1.051,
                                       Others = 1.372)


#Fryer's model with full control added as control
 fryer_full_control <- data.frame(Model = "(+) Precinct, Year", WhiteMean = "", Black = 1.178, Hispanic = 1.122, Asian = 0.953,
                                  Others = 1.060)


#creating a dataframe with Fryer's results
 fryer_results <- bind_rows(fryer_no_control, fryer_civilian_demographics,
                            fryer_encounter_characteristics,fryer_civilian_behavior, fryer_full_control)

 
 
#assigning new row names to the "fryer_results " data frame
 rownames(fryer_results ) <- c("(a)", "(b)",
                            "(c)",
                            "(d)", "(e)")

```

## Display Fryer's Results

```{r fryer-results}
kable(fryer_results, format = "html", booktabs = T, caption = "FRYER'S TABLE") %>%
  kable_styling(latex_options = c("striped", "hold_position"),
                full_width = F)
```

## Display Our Results

```{r our-results}
kable(our_results, format = "html", booktabs = T, caption = "OUR TABLE") %>%
  kable_styling(latex_options = c("striped", "hold_position"),
                full_width = F)
```
